Overriding and New Demo
This example shows when to use new or override modifier in derived class method.
Reference : https://msdn.microsoft.com/en-us/library/ms173153.aspx
1.  The override modifier extends the base class method, and the new modifier hides it. 
2. The new keyword preserves the relationships that produce that output, but it suppresses the warning. The variables that have type BaseClass continue to access the members of BaseClass, and the variable that has type DerivedClass continues to access members in DerivedClass first, and then to consider members inherited from BaseClass.
To suppress the warning, add the new modifier to the definition of Method2 in DerivedClass, as shown in the following code. The modifier can be added before or after public.